Move find-tweets.ts to content-twitter/lib
