+++ src/content-twitter.tsx
@@ -3,10 +3,10 @@ import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import browser from 'webextension-polyfill';
 import { CopyButton } from './content-twitter/component/copy-button';
+import { findTweets } from './content-twitter/lib/find-tweets';
 import { touchAction, updateAction } from './content-twitter/state';
 import { store } from './content-twitter/store';
 import { mutationRecordInfo } from './lib/dom';
-import { findTweets } from './lib/find-tweets';
 import { logger } from './lib/logger';
 import { TweetCopyResponseMessage } from './lib/message';
 import { storage } from './lib/storage';
