+++ src/content-twitter/lib/find-tweets.ts
@@ -1,6 +1,6 @@
-import { getNode, isElement } from './dom';
-import { Logger, logger as defaultLogger } from './logger';
-import { TweetID } from './tweet';
+import { getNode, isElement } from '../../lib/dom';
+import { Logger, logger as defaultLogger } from '../../lib/logger';
+import { TweetID } from '../../lib/tweet';
 
 export interface FindTweetResult {
   tweetID: TweetID;
